Intercept PATCH response to include updated frame's thumbnail info

+++ flipbooks/api/views.py
@@ -1,4 +1,5 @@
 import os
+from django.conf import settings
 from rest_framework import generics
 from rest_framework.parsers import FormParser,MultiPartParser,FileUploadParser
 
@@ -13,6 +14,14 @@ from ..models import (
     Strip,
     Frame
 )
+
+# ---------------------------------------------
+# This is moved to the end to 
+# prevent import timing conflict
+# ---------------------------------------------
+#custom helper functions 
+#from .helpermodule import helpers
+# ---------------------------------------------
 
 
 class FlipbookAPIListView(generics.ListAPIView):
@@ -111,8 +120,8 @@ class FrameUpdateAPIView(generics.UpdateAPIView):
     # def put(self, request, *args, **kwargs):
     #      
         
-    ''' Old image and thumbnails are deleted upon DELETE request.
-        Check the signal receivers in models.py '''
+    # Old image and thumbnails are deleted upon DELETE request.
+    # Check the signal receivers in models.py 
     def partial_update(self, request, *args, **kwargs):
         print("------------- partial update [PATCH] --------------")
         print("for frame id#{}".format(kwargs['pk']) )
@@ -132,11 +141,60 @@ class FrameUpdateAPIView(generics.UpdateAPIView):
                 # frame._old_image_paths = image_paths
  
                 # Remove old images
-                if frame.frame_image: 
+                print("======= PATCH: removing old images =========")
+                if frame and frame.frame_image: 
+                    print("does this frame has thumbnails?:")
+                    for thumbnail in frame.frame_image.get_thumbnails():
+                        print(thumbnail.path)
+                
                     frame.frame_image.delete_thumbnails() 
                     image_path = frame.frame_image.path
                     if image_path and os.path.isfile(image_path):
                         os.remove(image_path)
-             
-        return super(FrameUpdateAPIView, self).partial_update(request, *args, **kwargs)
-     
+
+
+        # intercept!!
+        resp = super(FrameUpdateAPIView, self).partial_update(request, *args, **kwargs)
+        
+        
+        print("")
+        print("========= PATCH response ==========")
+        #print("Responce data: {}".format(resp.data))
+        
+        #resp.data.update({"thumbnail": 1})
+        
+        print("thumbnails on this object:")
+        
+        frame = Frame.objects.filter(pk=kwargs['pk'])[0] # <- I do have to re-retrieve
+        
+        frame_thumbnails = {}
+        for thumbnail in frame.frame_image.get_thumbnails():
+            # print(dir(thumbnail))
+            print("url: {}".format(thumbnail.url))
+            print("dimension: {}".format(thumbnail._get_image_dimensions()))
+            # print("__dict__: {}".format(dir(thumbnail)))
+            
+            # can't get alias?
+            # have to sort them into alias before using it
+            matched_alias = thumbnailer_helpers.get_alias_dict(
+                thumbnail.url, 
+                thumbnail._get_image_dimensions(),
+                settings.THUMBNAIL_ALIASES['']
+                )
+            if matched_alias:
+                frame_thumbnails.update(matched_alias)
+                
+            
+        # inject frame_thumbnails into the response
+        print("frame_thumbnails: {}".format(frame_thumbnails))
+        resp["frame_thumbnails"] = frame_thumbnails
+        
+        print("===================================")
+        print("")
+        
+        return resp
+        
+        
+        
+# custom helper functions 
+from ..helpermodule import thumbnailer_helpers
